Added tensor view and iterator tags

+++ src/core/include/TensorII/Types.h
@@ -9,8 +9,8 @@ namespace TensorII::Core{
 
     using tensorDimension = long;  // 2^32 = 4G, probably don't need larger
     using tensorIndex = tensorDimension;
-    using tensorRank = size_t;
-    using tensorSize = size_t;
+    using tensorRank = unsigned long;
+    using tensorSize = unsigned long;
 
     struct from_range_t { explicit from_range_t() = default; };
     constexpr from_range_t from_range = from_range_t{};
